// src/main/resources/webapp/src/app/production/common/model/yieldParam.ts
export class YieldParam {
    shopName: string;
    productionType: string;
    productCategory: string;
    productCategorys: string[];
    operationName: string;
    datatype: string;
    processtype: string;
    linename: string;
    linetype: string;
    starttime: string;
    endtime: string;
    date: string;
    lineNames: string[];
    defectcode: string;
    datatype1: string;
    deviceType: string;
}